requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==2.3.2
pillow==11.3.0
matplotlib==3.10.5
tqdm==4.67.1
scikit-image==0.25.2
scipy==1.16.1
